Relocate UI constants
